pyproject.toml
[tool.poetry]
name = "nonebot-plugin-l4d2-server"
version = "0.4.0"
description = "gw2 plugin for NoneBot"
authors = ["Agnes_Digital <Z735803792@163.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Umamusume-Agnes-Digital/nonebot_plugin_gw2"
repository = "https://github.com/Umamusume-Agnes-Digital/nonebot_plugin_gw2"
keywords = ["game", "gw2", "nonebot2", "plugin"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Operating System :: OS Independent",
]
include = [
    "LICENSE","README.md"
]

[tool.poetry.dependencies]
python = "^3.7"
nonebot2 = "^2.0.0rc3"
nonebot-adapter-onebot = ">=2.1.5"
asyncio = ">=3.4.3"
httpx = "^0.23.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
